Split Base32768Test
