+++ Base32768.Test/Base32768Test.cs
@@ -1,56 +1,11 @@
 ﻿using System;
-using System.Text;
 using FluentAssertions;
 using Xunit;
 
 namespace Kzrnm.Convert.Base32768
 {
-    public class Base32768Test
+    public class Base32768Test : Base32768TestBase
     {
-        public static TheoryData Simple_Data = new TheoryData<string, byte[]>
-        {
-            {
-                "ꡟ",
-                new byte[]{255}
-            },
-            {
-                "ꡟꡟꡟꡟꡟꡟꡟꡟꡟ",
-                new byte[]{
-                    255, 255, 255, 255,
-                    255, 255, 255, 255,
-                    255, 255, 255, 255,
-                    255, 255, 255, 255
-                }
-            },
-            {
-                "ݠ暠䙠㙐▨ᖄቢႡဟ",
-                new byte[]{
-                    1, 1, 1, 1,
-                    1, 1, 1, 1,
-                    1, 1, 1, 1,
-                    1, 1, 1, 1
-                }
-            },
-            {
-                "ݠ暠䙠㙐▨ᖄቢႡݠʟ",
-                new byte[]{
-                    1, 1, 1, 1,
-                    1, 1, 1, 1,
-                    1, 1, 1, 1,
-                    1, 1, 1, 1, 1
-                }
-            },
-            {
-                "ݠ暠䙠㙐▨ᖄቢႡݠ曟",
-                new byte[]{
-                    1, 1, 1, 1,
-                    1, 1, 1, 1,
-                    1, 1, 1, 1,
-                    1, 1, 1, 1, 1, 1
-                }
-            }
-        };
-
         [Theory]
         [MemberData(nameof(Simple_Data))]
         public void Simple(string str, byte[] bytes)
@@ -66,7 +21,7 @@ namespace Kzrnm.Convert.Base32768
             Base32768.Encode(bytes.ToStream()).Should().Be(str);
         }
 
-#if NET5_0
+#if NETCOREAPP3_1_OR_GREATER
         [Theory]
         [MemberData(nameof(Simple_Data))]
         public void SimpleSpan(string str, byte[] bytes)
@@ -75,18 +30,6 @@ namespace Kzrnm.Convert.Base32768
             Base32768.Decode(str.AsSpan()).Should().Equal(bytes);
         }
 #endif
-        public static TheoryData EnumerateRandomBytes()
-        {
-            var rnd = new Random();
-            var theoryData = new TheoryData<byte[]>();
-            for (int i = 0; i < 100; i++)
-            {
-                var bytes = new byte[rnd.Next(1, 1000)];
-                rnd.NextBytes(bytes);
-                theoryData.Add(bytes);
-            }
-            return theoryData;
-        }
 
 #if !DEBUG
         [Theory]
@@ -98,26 +41,10 @@ namespace Kzrnm.Convert.Base32768
 
             Base32768.Encode(bytes.ToStream()).Should().Be(str);
             Base32768.Decode(str).Should().Equal(bytes);
-#if NET5_0
+#if NETCOREAPP3_1_OR_GREATER
             Base32768.Encode(bytes.AsSpan()).Should().Be(str);
             Base32768.Decode(str.AsSpan()).Should().Equal(bytes);
 #endif
-        }
-
-        public static TheoryData EnumeratePairTestData()
-        {
-            var testData = TestUtil.TestData;
-            var theoryData = new TheoryData<string, byte[]>();
-            foreach (var (name, val) in testData)
-            {
-                if (!name.StartsWith("test-data/pairs"))
-                    continue;
-                if (!name.EndsWith(".txt"))
-                    continue;
-                var binName = System.IO.Path.ChangeExtension(name, "bin");
-                theoryData.Add(Encoding.UTF8.GetString(val), testData[binName]);
-            }
-            return theoryData;
         }
 
         [Theory]
@@ -126,7 +53,7 @@ namespace Kzrnm.Convert.Base32768
         {
             Base32768.Encode(bytes).Should().Be(str);
             Base32768.Decode(str).Should().Equal(bytes);
-#if NET5_0
+#if NETCOREAPP3_1_OR_GREATER
             Base32768.Encode(bytes.ToStream()).Should().Be(str);
             Base32768.Encode(bytes.AsSpan()).Should().Be(str);
             Base32768.Decode(str.AsSpan()).Should().Equal(bytes);
@@ -134,27 +61,12 @@ namespace Kzrnm.Convert.Base32768
         }
 #endif
 
-        public static TheoryData EnumerateTestDataBad()
-        {
-            var testData = TestUtil.TestData;
-            var theoryData = new TheoryData<string>();
-            foreach (var (name, val) in testData)
-            {
-                if (!name.StartsWith("test-data/bad"))
-                    continue;
-                if (!name.EndsWith(".txt"))
-                    continue;
-                theoryData.Add(Encoding.UTF8.GetString(val));
-            }
-            return theoryData;
-        }
-
         [Theory]
         [MemberData(nameof(EnumerateTestDataBad))]
         public void Bad(string str)
         {
             str.Invoking(str => Base32768.Decode(str)).Should().Throw<FormatException>();
-#if NET5_0
+#if NETCOREAPP3_1_OR_GREATER
             str.Invoking(str => Base32768.Decode(str.AsSpan())).Should().Throw<FormatException>();
 #endif
         }
